fix(table_graph): move extract orchestrator to its own module
The extract orchestrator (formerly dossier::run) referenced sibling modules
discover and llm via crate::, which the test target couldn't resolve when
dossier.rs was #[path]-included. CI caught this; local cargo build passed
because we never re-ran cargo test after Task 10 added the function.

Move it to a dedicated extract.rs module; tests don't include extract.rs.
Also explicitly type the &[&str] slice arguments to avoid array-size
inference ambiguity surfaced by the cascading compile errors.
